Add a wrapper script for the config operation
It isn't very safe to have a script sitting in /sbin with no usage
information and were calling it with no args will immediately change the
system state.

Add a wrapper script with usage, GNU verison info, and accidental-call
safety. Calling the wrapper with 'config' or 'configure' as the command
will execute the same backing script.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>