Install a nilrt-snac-conflicts meta-package
Some parts of the SNAC configuration prohibit the installation of
certain packages like ni-auth or xserver. But these packages might
accidentally get reinstalled as dependencies of other package
installations.

As a protection against that happening, construct a meta-package called
nilrt-snac-conflicts, that contains no data, but conflicts with
prohibited packages. Install it to the system after uninstalling the
bad packages, to keep them from getting reinstalled.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>