OpkgHelper: defer update to install operation
Each time the OpkgHelper is initialized, it will attempt update the
package indexes. Not only is this technically a destructive operation,
but it is also not necessary unless doing an opkg installation
operation. As is, the nilrt-snac tool takes several seconds to perform
any operation, even if it is not installing any packages.

Defer the update operation out of __init__(), and prefer it to be
called at the start of a configure operation.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>