nilrt_snac: resolve the datadir path when assigned
The python __file__ variable evaluates to the relative path of the
python file *being executed*. Its value can be deferred for some reason
I don't understand. But when importing the datadir variable to other
modules, it will end up with a value relative to the importing module,
not the nilrt_snac module file.

So resolve() the path immediately in the nilrt_snac module, which fixes
the issue.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>