remote-test-firewall.sh: rewrite to use netns instead of multiple hosts
The present system-level test for the SNAC firewall has rather onerous system
configuration requirements, notably the need for a secondary network interface
between the Linux host and the SNAC DUT. Here, it is rewritten to run entirely
on different network namespaces on the SNAC DUT, allowing single-host,
zero-argument operation. The basic technique is borrowed from the firewalld test
suite.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>