Fix Initialization of auditd_config_file
Moved the initialization of `auditd_config_file` to occur after the potential installation of the auditd package. This change ensures that the auditd.conf is correctly populated. Previously, the file could be initialized as an empty string if it didn't exist, leading to the entire file being overwritten with an empty string.

Signed-off-by: Eli Engelhardt <eli.engelhardt@emerson.com>